refactor publish to use -merged sourtces if avaialble
