+++ src/stubber/commands/publish_cmd.py
@@ -4,7 +4,7 @@ from typing import List, Union
 
 import click
 from stubber.publish.database import get_database
-from stubber.publish.package import ALL_TYPES
+from stubber.publish.enums import ALL_TYPES
 from stubber.publish.publish import publish, publish_multiple
 from stubber.utils.config import CONFIG
 from stubber.utils.my_version import __version__
