+++ tests/publish/test_publish.py
@@ -4,7 +4,7 @@ import pytest
 from mock import MagicMock
 from pytest_mock import MockerFixture
 from stubber.publish.database import get_database
-from stubber.publish.package import ALL_TYPES, COMBO_STUBS, CORE_STUBS, DOC_STUBS
+from stubber.publish.enums import ALL_TYPES, COMBO_STUBS, CORE_STUBS, DOC_STUBS
 from stubber.publish.publish import publish, publish_multiple, publish_one
 from stubber.utils.config import CONFIG, readconfig
